N21-1337 removes deprecated ctl versioning (#4927)
* N21-1337 removes the deprecated feature flags: FEATURE_CTL_CONTEXT_CONFIGURATION_ENABLED and FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED
