+++ apps/server/src/modules/server/api/test/server.api.spec.ts
@@ -49,7 +49,6 @@ describe('Server Controller (API)', () => {
 				'FEATURE_CONSENT_NECESSARY',
 				'FEATURE_COPY_SERVICE_ENABLED',
 				'FEATURE_COURSE_SHARE',
-				'FEATURE_CTL_CONTEXT_CONFIGURATION_ENABLED',
 				'FEATURE_CTL_TOOLS_COPY_ENABLED',
 				'FEATURE_CTL_TOOLS_TAB_ENABLED',
 				'FEATURE_ENABLE_LDAP_SYNC_DURING_MIGRATION',
