+++ apps/server/src/modules/tool/common/service/common-tool.service.spec.ts
@@ -1,16 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import {
-	contextExternalToolFactory,
-	externalToolFactory,
-	schoolExternalToolFactory,
-	toolConfigurationStatusFactory,
-} from '@shared/testing';
-import { CommonToolService } from './common-tool.service';
+import { externalToolFactory } from '@shared/testing';
 import { ExternalTool } from '../../external-tool/domain';
-import { SchoolExternalTool } from '../../school-external-tool/domain';
 import { ToolContextType } from '../enum';
-import { ContextExternalToolConfigurationStatus } from '../domain';
-import { ContextExternalTool } from '../../context-external-tool/domain';
+import { CommonToolService } from './common-tool.service';
 
 describe('CommonToolService', () => {
 	let module: TestingModule;
@@ -26,298 +18,6 @@ describe('CommonToolService', () => {
 
 	afterAll(async () => {
 		await module.close();
-	});
-
-	describe('determineToolConfigurationStatus', () => {
-		describe('when all versions are equal', () => {
-			const setup = () => {
-				const externalTool: ExternalTool = externalToolFactory.buildWithId({ version: 0 });
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.buildWithId({ toolVersion: 0 });
-				const contextExternalTool: ContextExternalTool = contextExternalToolFactory.buildWithId({ toolVersion: 0 });
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should return a configuration status with latest true', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				const result: ContextExternalToolConfigurationStatus = service.determineToolConfigurationStatus(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-
-				expect(result).toEqual(
-					toolConfigurationStatusFactory.build({
-						isOutdatedOnScopeContext: false,
-						isOutdatedOnScopeSchool: false,
-					})
-				);
-			});
-		});
-
-		describe('when externalTool version is greater than schoolExternalTool version', () => {
-			const setup = () => {
-				const externalTool: ExternalTool = externalToolFactory.buildWithId({ version: 1 });
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.buildWithId({ toolVersion: 0 });
-				const contextExternalTool: ContextExternalTool = contextExternalToolFactory.buildWithId({ toolVersion: 0 });
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should return outdated status for school level', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				const result: ContextExternalToolConfigurationStatus = service.determineToolConfigurationStatus(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-
-				expect(result).toEqual(
-					toolConfigurationStatusFactory.build({
-						isOutdatedOnScopeContext: true,
-						isOutdatedOnScopeSchool: true,
-						isDeactivated: false,
-					})
-				);
-			});
-		});
-
-		describe('when schoolExternalTool version is greater than contextExternalTool version', () => {
-			const setup = () => {
-				const externalTool: ExternalTool = externalToolFactory.buildWithId({ version: 1 });
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.buildWithId({ toolVersion: 1 });
-				const contextExternalTool: ContextExternalTool = contextExternalToolFactory.buildWithId({ toolVersion: 0 });
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should return outdated status for context level', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				const result: ContextExternalToolConfigurationStatus = service.determineToolConfigurationStatus(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-
-				expect(result).toEqual(
-					toolConfigurationStatusFactory.build({
-						isOutdatedOnScopeContext: true,
-						isOutdatedOnScopeSchool: true,
-						isDeactivated: false,
-					})
-				);
-			});
-		});
-
-		describe('when externalTool version is greater than contextExternalTool version', () => {
-			const setup = () => {
-				const externalTool: ExternalTool = externalToolFactory.buildWithId({ version: 1 });
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.buildWithId({ toolVersion: 1 });
-				const contextExternalTool: ContextExternalTool = contextExternalToolFactory.buildWithId({ toolVersion: 0 });
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should return outdated status for context and school level', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				const result: ContextExternalToolConfigurationStatus = service.determineToolConfigurationStatus(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-
-				expect(result).toEqual(
-					toolConfigurationStatusFactory.build({
-						isOutdatedOnScopeContext: true,
-						isOutdatedOnScopeSchool: true,
-					})
-				);
-			});
-		});
-
-		describe('when contextExternalTool version is greater than schoolExternalTool version', () => {
-			const setup = () => {
-				const externalTool: ExternalTool = externalToolFactory.buildWithId({ version: 1 });
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.buildWithId({ toolVersion: 1 });
-				const contextExternalTool: ContextExternalTool = contextExternalToolFactory.buildWithId({ toolVersion: 2 });
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should return a configuration status with latest true', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				const result: ContextExternalToolConfigurationStatus = service.determineToolConfigurationStatus(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-
-				expect(result).toEqual(
-					toolConfigurationStatusFactory.build({
-						isOutdatedOnScopeContext: false,
-						isOutdatedOnScopeSchool: false,
-					})
-				);
-			});
-		});
-
-		describe('when contextExternalTool version is greater than externalTool version', () => {
-			const setup = () => {
-				const externalTool: ExternalTool = externalToolFactory.buildWithId({ version: 1 });
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.buildWithId({ toolVersion: 1 });
-				const contextExternalTool: ContextExternalTool = contextExternalToolFactory.buildWithId({ toolVersion: 2 });
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should return a configuration status with latest true', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				const result: ContextExternalToolConfigurationStatus = service.determineToolConfigurationStatus(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-
-				expect(result).toEqual(
-					toolConfigurationStatusFactory.build({
-						isOutdatedOnScopeContext: false,
-						isOutdatedOnScopeSchool: false,
-						isDeactivated: false,
-					})
-				);
-			});
-		});
-
-		describe('when schoolExternalTool version is greater than externalTool version', () => {
-			const setup = () => {
-				const externalTool: ExternalTool = externalToolFactory.buildWithId({ version: 1 });
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.buildWithId({ toolVersion: 2 });
-				const contextExternalTool: ContextExternalTool = contextExternalToolFactory.buildWithId({ toolVersion: 2 });
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should return a configuration status with latest true', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				const result: ContextExternalToolConfigurationStatus = service.determineToolConfigurationStatus(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-
-				expect(result).toEqual(
-					toolConfigurationStatusFactory.build({
-						isOutdatedOnScopeContext: false,
-						isOutdatedOnScopeSchool: false,
-						isDeactivated: false,
-					})
-				);
-			});
-		});
-
-		describe('when schoolExternalTool is deactivated', () => {
-			const setup = () => {
-				const externalTool: ExternalTool = externalToolFactory.buildWithId({ version: 1 });
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.buildWithId({
-					toolVersion: 2,
-					status: { isDeactivated: true },
-				});
-				const contextExternalTool: ContextExternalTool = contextExternalToolFactory.buildWithId({ toolVersion: 2 });
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should return a configuration status with deactivated true', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				const result: ContextExternalToolConfigurationStatus = service.determineToolConfigurationStatus(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-
-				expect(result).toEqual(
-					toolConfigurationStatusFactory.build({
-						isOutdatedOnScopeContext: false,
-						isOutdatedOnScopeSchool: false,
-						isDeactivated: true,
-					})
-				);
-			});
-		});
-
-		describe('when externalTool is deactivated', () => {
-			const setup = () => {
-				const externalTool: ExternalTool = externalToolFactory.buildWithId({ version: 1, isDeactivated: true });
-				const schoolExternalTool: SchoolExternalTool = schoolExternalToolFactory.buildWithId({
-					toolVersion: 2,
-				});
-				const contextExternalTool: ContextExternalTool = contextExternalToolFactory.buildWithId({ toolVersion: 2 });
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should return a configuration status with deactivated true', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				const result: ContextExternalToolConfigurationStatus = service.determineToolConfigurationStatus(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-
-				expect(result).toEqual(
-					toolConfigurationStatusFactory.build({
-						isOutdatedOnScopeContext: false,
-						isOutdatedOnScopeSchool: false,
-						isDeactivated: true,
-					})
-				);
-			});
-		});
 	});
 
 	describe('isContextRestricted', () => {
