+++ apps/server/src/modules/tool/context-external-tool/context-external-tool.module.ts
@@ -6,7 +6,7 @@ import { SchoolExternalToolModule } from '../school-external-tool';
 import { ToolConfigModule } from '../tool-config.module';
 import { ContextExternalToolAuthorizableService, ContextExternalToolService, ToolReferenceService } from './service';
 import { ContextExternalToolValidationService } from './service/context-external-tool-validation.service';
-import { ToolVersionService } from './service/tool-version-service';
+import { ToolConfigurationStatusService } from './service/tool-configuration-status.service';
 
 @Module({
 	imports: [
@@ -21,14 +21,14 @@ import { ToolVersionService } from './service/tool-version-service';
 		ContextExternalToolValidationService,
 		ContextExternalToolAuthorizableService,
 		ToolReferenceService,
-		ToolVersionService,
+		ToolConfigurationStatusService,
 	],
 	exports: [
 		ContextExternalToolService,
 		ContextExternalToolValidationService,
 		ContextExternalToolAuthorizableService,
 		ToolReferenceService,
-		ToolVersionService,
+		ToolConfigurationStatusService,
 	],
 })
 export class ContextExternalToolModule {}
