+++ apps/server/src/modules/tool/context-external-tool/service/tool-configuration-status.service.spec.ts
@@ -12,43 +12,28 @@ import {
 	ToolParameterDuplicateLoggableException,
 	ToolParameterValueMissingLoggableException,
 } from '../../common/domain';
-import { CommonToolService, CommonToolValidationService } from '../../common/service';
-import { IToolFeatures, ToolFeatures } from '../../tool-config';
-import { ToolVersionService } from './tool-version-service';
+import { CommonToolValidationService } from '../../common/service';
+import { ToolConfigurationStatusService } from './tool-configuration-status.service';
 
-describe('ToolVersionService', () => {
+describe(ToolConfigurationStatusService.name, () => {
 	let module: TestingModule;
-	let service: ToolVersionService;
+	let service: ToolConfigurationStatusService;
 
 	let commonToolValidationService: DeepMocked<CommonToolValidationService>;
-	let commonToolService: DeepMocked<CommonToolService>;
-	let toolFeatures: DeepMocked<IToolFeatures>;
 
 	beforeAll(async () => {
 		module = await Test.createTestingModule({
 			providers: [
-				ToolVersionService,
+				ToolConfigurationStatusService,
 				{
 					provide: CommonToolValidationService,
 					useValue: createMock<CommonToolValidationService>(),
 				},
-				{
-					provide: CommonToolService,
-					useValue: createMock<CommonToolService>(),
-				},
-				{
-					provide: ToolFeatures,
-					useValue: {
-						toolStatusWithoutVersions: false,
-					},
-				},
 			],
 		}).compile();
 
-		service = module.get(ToolVersionService);
+		service = module.get(ToolConfigurationStatusService);
 		commonToolValidationService = module.get(CommonToolValidationService);
-		commonToolService = module.get(CommonToolService);
-		toolFeatures = module.get(ToolFeatures);
 	});
 
 	afterAll(async () => {
@@ -60,7 +45,7 @@ describe('ToolVersionService', () => {
 	});
 
 	describe('determineToolConfigurationStatus', () => {
-		describe('when FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED is false', () => {
+		describe('when validation runs through', () => {
 			const setup = () => {
 				const externalTool = externalToolFactory.buildWithId();
 				const schoolExternalTool = schoolExternalToolFactory.buildWithId({
@@ -69,40 +54,6 @@ describe('ToolVersionService', () => {
 				const contextExternalTool = contextExternalToolFactory
 					.withSchoolExternalToolRef(schoolExternalTool.id as string)
 					.buildWithId();
-
-				toolFeatures.toolStatusWithoutVersions = false;
-
-				return {
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool,
-				};
-			};
-
-			it('should call CommonToolService', () => {
-				const { externalTool, schoolExternalTool, contextExternalTool } = setup();
-
-				service.determineToolConfigurationStatus(externalTool, schoolExternalTool, contextExternalTool);
-
-				expect(commonToolService.determineToolConfigurationStatus).toHaveBeenCalledWith(
-					externalTool,
-					schoolExternalTool,
-					contextExternalTool
-				);
-			});
-		});
-
-		describe('when FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED is true and validation runs through', () => {
-			const setup = () => {
-				const externalTool = externalToolFactory.buildWithId();
-				const schoolExternalTool = schoolExternalToolFactory.buildWithId({
-					toolId: externalTool.id as string,
-				});
-				const contextExternalTool = contextExternalToolFactory
-					.withSchoolExternalToolRef(schoolExternalTool.id as string)
-					.buildWithId();
-
-				toolFeatures.toolStatusWithoutVersions = true;
 
 				commonToolValidationService.validateParameters.mockReturnValue([]);
 
@@ -147,7 +98,7 @@ describe('ToolVersionService', () => {
 			});
 		});
 
-		describe('when FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED is true and validation of SchoolExternalTool throws an error', () => {
+		describe('when validation of SchoolExternalTool throws an error', () => {
 			const setup = () => {
 				const externalTool = externalToolFactory.buildWithId();
 				const schoolExternalTool = schoolExternalToolFactory.buildWithId({
@@ -156,8 +107,6 @@ describe('ToolVersionService', () => {
 				const contextExternalTool = contextExternalToolFactory
 					.withSchoolExternalToolRef(schoolExternalTool.id as string)
 					.buildWithId();
-
-				toolFeatures.toolStatusWithoutVersions = true;
 
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
 				commonToolValidationService.validateParameters.mockReturnValueOnce([]);
@@ -203,7 +152,7 @@ describe('ToolVersionService', () => {
 			});
 		});
 
-		describe('when FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED is true and validation of ContextExternalTool throws an error', () => {
+		describe('when validation of ContextExternalTool throws an error', () => {
 			const setup = () => {
 				const externalTool = externalToolFactory.buildWithId();
 				const schoolExternalTool = schoolExternalToolFactory.buildWithId({
@@ -212,8 +161,6 @@ describe('ToolVersionService', () => {
 				const contextExternalTool = contextExternalToolFactory
 					.withSchoolExternalToolRef(schoolExternalTool.id as string)
 					.buildWithId();
-
-				toolFeatures.toolStatusWithoutVersions = true;
 
 				commonToolValidationService.validateParameters.mockReturnValueOnce([]);
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
@@ -259,7 +206,7 @@ describe('ToolVersionService', () => {
 			});
 		});
 
-		describe('when FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED is true and validation of SchoolExternalTool and  ContextExternalTool throws an error', () => {
+		describe('when validation of SchoolExternalTool and  ContextExternalTool throws an error', () => {
 			const setup = () => {
 				const externalTool = externalToolFactory.buildWithId();
 				const schoolExternalTool = schoolExternalToolFactory.buildWithId({
@@ -268,8 +215,6 @@ describe('ToolVersionService', () => {
 				const contextExternalTool = contextExternalToolFactory
 					.withSchoolExternalToolRef(schoolExternalTool.id as string)
 					.buildWithId();
-
-				toolFeatures.toolStatusWithoutVersions = true;
 
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
@@ -315,7 +260,7 @@ describe('ToolVersionService', () => {
 			});
 		});
 
-		describe('when FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED is true and validation of ContextExternalTool throws at least 1 missing value errors', () => {
+		describe('when validation of ContextExternalTool throws at least 1 missing value errors', () => {
 			const setup = () => {
 				const customParameter = customParameterFactory.build();
 				const externalTool = externalToolFactory.buildWithId({ parameters: [customParameter] });
@@ -325,8 +270,6 @@ describe('ToolVersionService', () => {
 				const contextExternalTool = contextExternalToolFactory
 					.withSchoolExternalToolRef(schoolExternalTool.id as string)
 					.buildWithId();
-
-				toolFeatures.toolStatusWithoutVersions = true;
 
 				commonToolValidationService.validateParameters.mockReturnValueOnce([]);
 				commonToolValidationService.validateParameters.mockReturnValueOnce([
@@ -359,7 +302,7 @@ describe('ToolVersionService', () => {
 			});
 		});
 
-		describe('when FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED is true and  SchoolExternalTool is deactivated', () => {
+		describe('when SchoolExternalTool is deactivated', () => {
 			const setup = () => {
 				const externalTool = externalToolFactory.buildWithId();
 				const schoolExternalTool = schoolExternalToolFactory.buildWithId({
@@ -369,8 +312,6 @@ describe('ToolVersionService', () => {
 				const contextExternalTool = contextExternalToolFactory
 					.withSchoolExternalToolRef(schoolExternalTool.id as string)
 					.buildWithId();
-
-				toolFeatures.toolStatusWithoutVersions = true;
 
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
@@ -400,7 +341,7 @@ describe('ToolVersionService', () => {
 			});
 		});
 
-		describe('when FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED is true and  externalTool is deactivated', () => {
+		describe('when externalTool is deactivated', () => {
 			const setup = () => {
 				const externalTool = externalToolFactory.buildWithId({ isDeactivated: true });
 				const schoolExternalTool = schoolExternalToolFactory.buildWithId({
@@ -409,8 +350,6 @@ describe('ToolVersionService', () => {
 				const contextExternalTool = contextExternalToolFactory
 					.withSchoolExternalToolRef(schoolExternalTool.id as string)
 					.buildWithId();
-
-				toolFeatures.toolStatusWithoutVersions = true;
 
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
@@ -440,7 +379,7 @@ describe('ToolVersionService', () => {
 			});
 		});
 
-		describe('when FEATURE_COMPUTE_TOOL_STATUS_WITHOUT_VERSIONS_ENABLED is true,  externalTool and schoolExternalTool are not deactivated', () => {
+		describe('when externalTool and schoolExternalTool are not deactivated', () => {
 			const setup = () => {
 				const externalTool = externalToolFactory.buildWithId();
 				const schoolExternalTool = schoolExternalToolFactory.buildWithId({
@@ -449,8 +388,6 @@ describe('ToolVersionService', () => {
 				const contextExternalTool = contextExternalToolFactory
 					.withSchoolExternalToolRef(schoolExternalTool.id as string)
 					.buildWithId();
-
-				toolFeatures.toolStatusWithoutVersions = true;
 
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
 				commonToolValidationService.validateParameters.mockReturnValueOnce([new ValidationError('')]);
