+++ apps/server/src/modules/tool/context-external-tool/service/tool-reference.service.spec.ts
@@ -12,7 +12,7 @@ import { SchoolExternalToolService } from '../../school-external-tool/service';
 import { ToolReference } from '../domain';
 import { ContextExternalToolService } from './context-external-tool.service';
 import { ToolReferenceService } from './tool-reference.service';
-import { ToolVersionService } from './tool-version-service';
+import { ToolConfigurationStatusService } from './tool-configuration-status.service';
 
 describe('ToolReferenceService', () => {
 	let module: TestingModule;
@@ -21,7 +21,7 @@ describe('ToolReferenceService', () => {
 	let externalToolService: DeepMocked<ExternalToolService>;
 	let schoolExternalToolService: DeepMocked<SchoolExternalToolService>;
 	let contextExternalToolService: DeepMocked<ContextExternalToolService>;
-	let toolVersionService: DeepMocked<ToolVersionService>;
+	let toolVersionService: DeepMocked<ToolConfigurationStatusService>;
 	let externalToolLogoService: DeepMocked<ExternalToolLogoService>;
 
 	beforeAll(async () => {
@@ -41,8 +41,8 @@ describe('ToolReferenceService', () => {
 					useValue: createMock<ContextExternalToolService>(),
 				},
 				{
-					provide: ToolVersionService,
-					useValue: createMock<ToolVersionService>(),
+					provide: ToolConfigurationStatusService,
+					useValue: createMock<ToolConfigurationStatusService>(),
 				},
 				{
 					provide: ExternalToolLogoService,
@@ -55,7 +55,7 @@ describe('ToolReferenceService', () => {
 		externalToolService = module.get(ExternalToolService);
 		schoolExternalToolService = module.get(SchoolExternalToolService);
 		contextExternalToolService = module.get(ContextExternalToolService);
-		toolVersionService = module.get(ToolVersionService);
+		toolVersionService = module.get(ToolConfigurationStatusService);
 		externalToolLogoService = module.get(ExternalToolLogoService);
 	});
