+++ apps/server/src/modules/tool/school-external-tool/service/school-external-tool-validation.service.ts
@@ -1,25 +1,19 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { ValidationError } from '@shared/common';
 import { CommonToolValidationService } from '../../common/service';
 import { ExternalTool } from '../../external-tool/domain';
 import { ExternalToolService } from '../../external-tool/service';
-import { IToolFeatures, ToolFeatures } from '../../tool-config';
 import { SchoolExternalTool } from '../domain';
 
 @Injectable()
 export class SchoolExternalToolValidationService {
 	constructor(
 		private readonly externalToolService: ExternalToolService,
-		private readonly commonToolValidationService: CommonToolValidationService,
-		@Inject(ToolFeatures) private readonly toolFeatures: IToolFeatures
+		private readonly commonToolValidationService: CommonToolValidationService
 	) {}
 
 	async validate(schoolExternalTool: SchoolExternalTool): Promise<void> {
 		const loadedExternalTool: ExternalTool = await this.externalToolService.findById(schoolExternalTool.toolId);
-
-		if (!this.toolFeatures.toolStatusWithoutVersions) {
-			this.checkVersionMatch(schoolExternalTool.toolVersion, loadedExternalTool.version);
-		}
 
 		const errors: ValidationError[] = this.commonToolValidationService.validateParameters(
 			loadedExternalTool,
@@ -28,14 +22,6 @@ export class SchoolExternalToolValidationService {
 
 		if (errors.length) {
 			throw errors[0];
-		}
-	}
-
-	private checkVersionMatch(schoolExternalToolVersion: number, externalToolVersion: number): void {
-		if (schoolExternalToolVersion !== externalToolVersion) {
-			throw new ValidationError(
-				`tool_version_mismatch: The version ${schoolExternalToolVersion} of given schoolExternalTool does not match the externalTool version ${externalToolVersion}.`
-			);
 		}
 	}
 }
